captain_console/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
inflection==0.4.0
psycopg2==2.8.5
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
